+++ src/bondgraph/elements/hydraulics.py
@@ -6,21 +6,23 @@ from typing import Set
 
 class Element_ReliefValve(OnePortElement):
     def __init__(self, name: str, ro: Symbol, rc: Symbol, k: Symbol, d: Symbol):
-        super().__init__(name, 'Rv')
+        super().__init__(name)
         self.ro = ro
         self.rc = rc
         self.d = d
         self.k = k
 
     def equations(self, effort: Function, flow: Function, time: Symbol):
-        return [Equality(
-            flow(time),
-            effort(time)
-            / (
-                self.rc * (0.5 - 0.5 * tanh(self.k * (effort(time) - self.d)))
-                + self.ro * (0.5 + 0.5 * tanh(self.k * (effort(time) - self.d)))
-            ),
-        )]
+        return [
+            Equality(
+                flow(time),  # type: ignore
+                effort(time)  # type: ignore
+                / (
+                    self.rc * (0.5 - 0.5 * tanh(self.k * (effort(time) - self.d)))  # type: ignore
+                    + self.ro * (0.5 + 0.5 * tanh(self.k * (effort(time) - self.d)))  # type: ignore
+                ),
+            )
+        ]
 
     @staticmethod
     def causality_policy():
@@ -28,3 +30,6 @@ class Element_ReliefValve(OnePortElement):
 
     def parameter_symbols(self) -> Set[Symbol]:
         return {self.ro, self.rc, self.d, self.k}
+
+    def visualization_label(self) -> str:
+        return "Rv"